requirements.txt
Django==1.4.1
django-fiber==0.9.7.1
MySQL-python==1.2.3
South==0.7.5
django-guardian==1.0.4
